requirements.txt

opencv-python
numpy
astral
requests
schedule
onvif-zeep
openai
python-dotenv